+++ option.go
@@ -1,5 +1,10 @@
 package options
 
-// Option should be provided by wrapper function as value setter.
-// It will be called when passed to Options.Resolve.
-type Option func(Options)
+type (
+	// Options is collection of Option value setters.
+	Options []Option
+
+	// Option should be provided by wrapper function as value setter.
+	// It will be called when passed to Resolver.Resolve.
+	Option func(Resolver)
+)