+++ options.go
@@ -5,15 +5,15 @@ import (
 	"fmt"
 )
 
-// Resolve will create Options instance filled with values based on received Option modificators.
-func Resolve(opts ...Option) Options {
-	return newOptions().Resolve(opts...)
+// Resolve will create Resolver instance filled with values based on received Option modificators.
+func Resolve(opts Options) Resolver {
+	return newResolver().Resolve(opts)
 }
 
-// Read will attempt to acquire value associated with provided key from received Options set.
+// Read will attempt to acquire value associated with provided key from received Resolver.
 // If operation fails due to value being missing it will respond with ErrNotFound.
 // If operation fails due to mismatch type between provided value and stored value then it will respond with ErrTypeMismatch.
-func Read[T any](options Options, key OptionKey) (T, error) {
+func Read[T any](options Resolver, key OptionKey) (T, error) {
 	var t T
 	val, typeName, err := options.getRegistry().get(key)
 	if err != nil {
@@ -34,9 +34,9 @@ func Read[T any](options Options, key OptionKey) (T, error) {
 }
 
 // ReadOrDefault acts exactly the same as Read with one difference being the option for providing default value if
-// there is none set in received Options set under provided key.
+// there is none set in received Resolver under provided key.
 // Attempt to acquire value can still fail with ErrTypeMismatch is types do not match.
-func ReadOrDefault[T any](options Options, key OptionKey, defaultValue T) (T, error) {
+func ReadOrDefault[T any](options Resolver, key OptionKey, defaultValue T) (T, error) {
 	val, err := Read[T](options, key)
 	if err != nil {
 		if (any(defaultValue) == nil && errors.Is(err, ErrNilValue)) || errors.Is(err, ErrNotFound) {
@@ -52,7 +52,7 @@ func ReadOrDefault[T any](options Options, key OptionKey, defaultValue T) (T, er
 // ReadOrPanic acts exactly the same as Read, but any error occurring during attempt to acquire value
 // will result in panic instead of returning an error. This also means that number of return parameters
 // is reduced to only one.
-func ReadOrPanic[T any](options Options, key OptionKey) T {
+func ReadOrPanic[T any](options Resolver, key OptionKey) T {
 	val, err := Read[T](options, key)
 	if err != nil {
 		panic(err)
@@ -63,7 +63,7 @@ func ReadOrPanic[T any](options Options, key OptionKey) T {
 // ReadOrDefaultOrPanic acts exactly the same as ReadOrDefault, but any error occurring during attempt to acquire value
 // (except for ErrNotFound) will result in panic instead of returning an error. This also means that number of return parameters
 // is reduced to only one.
-func ReadOrDefaultOrPanic[T any](options Options, key OptionKey, defaultValue T) T {
+func ReadOrDefaultOrPanic[T any](options Resolver, key OptionKey, defaultValue T) T {
 	val, err := ReadOrDefault[T](options, key, defaultValue)
 	if err != nil {
 		panic(err)
@@ -71,9 +71,9 @@ func ReadOrDefaultOrPanic[T any](options Options, key OptionKey, defaultValue T)
 	return val
 }
 
-// Write will attempt to store supplied value and associate it with provided key in received Options set.
-// If operation fails due to key being already present within received Options set it will respond with ErrDuplicatedKey.
-func Write[T any](options Options, key OptionKey, val T) error {
+// Write will attempt to store supplied value and associate it with provided key in received Resolver.
+// If operation fails due to key being already present within received Resolver it will respond with ErrDuplicatedKey.
+func Write[T any](options Resolver, key OptionKey, val T) error {
 	if err := options.getRegistry().set(key, val); err != nil {
 		return fmt.Errorf("cannot write option '%s' to provided option set: %w", key.String(), err)
 	}
@@ -83,7 +83,7 @@ func Write[T any](options Options, key OptionKey, val T) error {
 // WriteOrPanic acts exactly the same as Write, but any error occurring during attempt to store value
 // will result in panic instead of returning an error. This also means there are no longer any return parameters
 // needed for this operation.
-func WriteOrPanic[T any](options Options, key OptionKey, val T) {
+func WriteOrPanic[T any](options Resolver, key OptionKey, val T) {
 	if err := Write[T](options, key, val); err != nil {
 		panic(err)
 	}